cloud_function/requirements.txt
functions-framework==3.*
Flask
google-cloud-aiplatform